User interface: User functionalty
added the user interface for the user
now the user can login, search for flight, filter flights, book a
flight, mangae bookings by viewing them, modify them or cancel them.

+++ AirportTicketBooking/Details/UserDetails.cs
@@ -36,7 +36,7 @@ public class UserDetails : IIndexed
     public void ModifyBooking(int bookingId, Classes newClass)
     {
         BookingStorage bookingStorage = BookingStorage.GetStorageInstance();
-        if (bookingStorage.FindBooking(bookingId) != null)
+        if (bookingStorage.FindBooking(bookingId) != null && BookingIds.Contains(bookingId))
         {
             var flightId = bookingStorage.FindBooking(bookingId).FlightId;
             bookingStorage.DeleteData(bookingId);
@@ -51,7 +51,7 @@ public class UserDetails : IIndexed
     public void CancelBooking(int bookingId)
     {
         BookingStorage bookingStorage = BookingStorage.GetStorageInstance();
-        if (bookingStorage.FindBooking(bookingId) != null)
+        if (bookingStorage.FindBooking(bookingId) != null && BookingIds.Contains(bookingId))
         {
             bookingStorage.DeleteData(bookingId);
             BookingIds.Remove(bookingId);
